fix: Resolve build errors - separate client/server questionnaire engine, fix lint and type errors
