+++ src/app/api/questionnaire/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getCurrentUser } from "@/lib/auth";
+import { getQuestionByStep } from "@/lib/questionnaire/engine-client";
 import {
-  getQuestionByStep,
   saveResponses,
   getResponses,
 } from "@/lib/questionnaire/engine";
